Barebone CfL
Basic CfL support is added. Luma coefficients are stored and the AC
coefficients are reused as prediction for Chroma.

This is only compatible with 4:2:0 and when the Luma transform size
is equivalent to the Chroma transform size, 4 Luma blocks are TF
together and the AC coefficients from the top left quadrant are used.
When the Chroma transform size is half of that of the Luma transform
size, only the AC coefficients from the top left corner are used. This
is equivalent to DCT-domain subsampling.

If PVQ determines that the AC coefficients are skipped, the AC
coefficients of the Luma Intra prediction are stored instead. To allow
the decoder to compensate for the eventuality of a late skip, the decoder
stores the AC coefficients of the Luma Intra prediction even when the
block is skipped.

Compared to removing Chroma Intra prediction, barebone CfL results in an
important rate increase:

  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
6.4350 |  3.6294 |  2.6868 |   6.7849 | 6.4159 |  6.5020 |     6.3626