Enable PVQ keyframe code for Intra Chroma planes
To use CfL on Intra Chroma planes, we need to enable the Daala's PVQ
keyframe code. This code removes gain prediction and zeros out the
predictions of skipped blocks.

Another feature is signaling a flipbit for CfL. This feature remains
disabled and  will be enable in a subsequent commit. Since HAAR DC
is not present in AV1, zeroing out the DC is disabled.

Change-Id: If3dce50ceef560c7058d3d188ec441cf9cdf3000

Compared to removing Chroma Intra prediction, enabling the PVQ
keyframe code results in the following BD-rate increase:

  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
 -0.02 |    0.24 |    0.04 |    -0.01 |  -0.01 |   -0.01 |       0.06