CfL performed when PVQ is called from Chroma
Prior versions of this prototype only applied CfL on the final call to
PVQ (i.e. right before coding the bitstream). CfL is now applied
everytime PVQ is invoked.

This allowed to remove the code that restricted CfL flipping to the
final call to PVQ.

However, this requires an extra PVQ encode operation after the optimal
Luma parameters have been determined but before Chroma is tested. We use
this extra encode to store the best Luma coeffs as they are needed to
predict Chroma coeffs.

CfL now supports cases, where the Chroma transform size is greather than
the Luma transform size. This is acheived by applying a TF merge to
increase the Luma transform size so it matches the Chroma transform
size. To support multiple incremental TF merges a double buffer system
is used.

When multiple Luma transforms are used, CfL require that these be all of
the same size. This is not a problem with the current implementation of
the AV1.

An important fix with the prior version of CfL is that the coeffs used
to predict Chroma are now the first AC values of Luma and not the
collocated values.

The buffer used to store the CfL prediction is now zeroed out when on
the first luma store of a partition. It was discovered that certain edge
condition caused stale memory reads. Zeroing out these value solves this
problem.

Change-Id: I05839d0af0d9254ab636cac0b8fcee35711d131c