CfL Zero out prediction buffer
We can't always garantee that the Luma values will be available to serve
as prediction for Chroma. One such example is on frame boundary, where
the range of subsampled Chroma coeff reach the outside of the frame. In
this situation, Luma is never called outside the frame and no prediciton
is available. Instead of using stale memory, we now use zeros.

Change-Id: I911a8ac02395985e73073c0466b7d9bff82ad6f2