Fast DC computation for DC_PRED
The value of the DC of an NXN DC_PRED block is 8*N*x, where x is the
value used for the DC_PRED. 8*N*x maps to the DC obtained with
aom_fdctNxN_1_c. However, it does not map to the DC obtained with
aom_fdctNxN_c. This is probably caused by rounding issues in
aom_fdctNxN_c. Because of this, using 8*N*x to compute the DC of a
DC_PRED used in PVQ causes the following regression in AV1

  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.1266 |  0.0056 | -0.0100 |   0.1836 | 0.1733 |  0.1642 |     0.0992
https://arewecompressedyet.com/?job=av1_master_pvq%402017-01-27T03%3A26%3A26.934Z&job=Fast_Transform_For_DCPRED%402017-01-27T13%3A32%3A38.574Z

Change-Id: I195d05b49c201634704a4b06270f4c436d26d0cf