Standalone computation of the DC
The 1/sqrt(2) scaling factor applied to the DC introduces rounding
error. In the case of a 2D DCT, the rounding error is compounded. The rounding
error should be only that of dividing by 2, since (1/sqrt(2))^2=1/2.

Since removing the scaling factor will change the resulting AC
coefficients, this patch computes the DC standalone and overwrites it.

This patch applies to the C implementation of the  DCT, and thus requires
hijacking the DCT to for the C implementation of the DCT.

The resulting BD-rates are

  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
  0.0250 | -0.0359 | -0.1205 |   0.1199 | 0.0303 |  0.0304 |     0.0132
https://arewecompressedyet.com/?job=av1_master_pvq%402017-01-27T03%3A26%3A26.934Z&job=DCT_DC%402017-01-31T15%3A48%3A20.608Z

Change-Id: I2a217fd740774255eee28cdb0b5a60f16387954f