Enabling Daala's PVQ CFL Code in AV1
The PVQ code in AV1 already include some CfL features that are disabled
by default. This patch enables some of them.

Gain prediction is now disabled when CfL is used. Gain scaling also
differs when CfL is used.

This patch does not enable CfL flipping as the signaling is not
available in the bitstream and that it would interfere with RDO. CfL
flipping will be enabled when CfL is added to RDO.

Change-Id: I99a74f9db2ca4b2fc181d0b97f062422dc4800db