Replace is_keyframe checks with skip_can_copy check in PVQ
Because of H\V pred and CfL, block level skip on keyframes in Daala
implied zeroing out the coeff instead of copying the prediction. In AV1,
block level skips in keyframes impli copying the prediction.

To better reflect the nature of AV1, the checks are done on
skip_can_copy instead of is_keyframe.

Change-Id: I7abcb5e905bacbcf5d464094057d27899f0e5f83